multi-presence: Add optional property
Add the optional property DbusProperty to specify the name
of the property to update. This allows to update Dbus properties
besides the default 'Present'.

This allows multi-presence to not only detect the presence of FRU
devices, but also configuration options like the existance of HBM
on Intel SPR CPUs that is advertise using the PKG_ID pins.

Change-Id: I5061526a01997209eee96e42c698782ba226aa26
Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>